Update necessary test files after refactoring
