+++ src/test/java/seedu/jarvis/model/consultation/MasteryCheckTest.java
@@ -7,6 +7,7 @@ import java.time.LocalDateTime;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.jarvis.model.masteryCheck.MasteryCheck;
 import seedu.jarvis.testutil.StudentBuilder;
